+++ src/main/java/kr/carz/savecar/dto/TwoYearlyRentDTO.java
@@ -17,9 +17,9 @@ public class TwoYearlyRentDTO {
     private String category2; // 중형차, 경차 등 큰 분류
     private String deposit;
     private String name;
-    private Float cost_for_20Tk;
-    private Float cost_for_30Tk;
-    private Float cost_for_40Tk;
+    private double cost_for_20Tk;
+    private double cost_for_30Tk;
+    private double cost_for_40Tk;
     private String cost_for_others;
     private String age_limit;
     private String cost_per_km;
